% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.r
\name{search_station}
\alias{search_station}
\title{Search Station}
\usage{
search_station(q, n_res = 10L, throttle = 5)
}
\arguments{
\item{q}{character string, query to search for station.}

\item{n_res}{number of results to return (integer or character).}

\item{throttle}{how many request per minute to perform.}
}
\value{
tibble, containing station name and id
}
\description{
Search the station id using a query, returns n_res options in a tibble
}
\examples{
search_station("Frankfurt")
}
